[NPU] talker support
