+++ sglang_omni/platforms/cuda.py
@@ -8,9 +8,24 @@ from sglang.srt.platforms.cuda import CudaDeviceMixin
 from sglang.srt.platforms.rocm import RocmDeviceMixin
 
 from sglang_omni.platforms.interface import OmniPlatform
+from sglang_omni.utils.misc import normalize_quantization
 
 if TYPE_CHECKING:
     from sglang_omni.pipeline.stage_workers import StageLaunchConfig
+
+
+def _is_h20_device() -> bool:
+    """True only on NVIDIA H20 (word-boundary match so "H200" isn't caught)."""
+    try:
+        import re
+
+        import torch
+
+        if not torch.cuda.is_available():
+            return False
+        return bool(re.search(r"\bH20\b", torch.cuda.get_device_name(0)))
+    except Exception:
+        return False
 
 
 class CUDAOmniPlatform(CudaDeviceMixin, OmniPlatform):
@@ -42,6 +57,123 @@ class CUDAOmniPlatform(CudaDeviceMixin, OmniPlatform):
             "SGLANG_ONE_VISIBLE_DEVICE_PER_PROCESS": "true",
             "SGLANG_ENABLE_TP_MEMORY_INBALANCE_CHECK": "false",
         }
+
+    def get_intra_node_transport(self) -> TransportKind:
+        from sglang_omni.comm.data_ref import TransportKind
+
+        return TransportKind.CUDA_IPC
+
+    def get_fused_qk_norm_rope(self):
+        from sgl_kernel import fused_qk_norm_rope
+
+        return fused_qk_norm_rope
+
+    def apply_model_worker_backend_policy(
+        self,
+        server_args: ServerArgs,
+        model_config: ModelConfig,
+        model_arch_override: str | None,
+    ) -> str | None:
+
+        effective_quantization = normalize_quantization(model_config.quantization)
+        server_quantization = normalize_quantization(server_args.quantization)
+        if server_quantization is not None:
+            effective_quantization = server_quantization
+
+        moe_runner_backend = server_args.moe_runner_backend
+        is_qwen3_omni_arch = model_arch_override in (
+            "Qwen3OmniTalker",
+            "Qwen3OmniThinkerForCausalLM",
+        )
+        if is_qwen3_omni_arch and server_args.ep_size != 1:
+            raise ValueError(
+                "Qwen3-Omni ModelWorker does not support expert parallelism; "
+                "use ep_size=1."
+            )
+        has_moe = _model_config_has_moe(model_config)
+        has_native_fp8_block_quant = _model_config_has_native_fp8_block_quant(
+            model_config
+        )
+
+        if (
+            model_arch_override == "Qwen3OmniTalker"
+            and effective_quantization is None
+            and moe_runner_backend == "auto"
+        ):
+            # Note:(Chenchen Hong) flashinfer_cutlass MoE deadlocks CUDA-graph
+            # capture on H20 (no H20 kernel coverage); triton captures cleanly there.
+            override_server_args(
+                server_args,
+                "sglang-omni-qwen3-backend-policy",
+                moe_runner_backend=(
+                    "triton" if _is_h20_device() else "flashinfer_cutlass"
+                ),
+            )
+            moe_runner_backend = server_args.moe_runner_backend
+
+        if (
+            is_qwen3_omni_arch
+            and effective_quantization == "fp8"
+            and has_moe
+            and moe_runner_backend == "auto"
+            and has_native_fp8_block_quant
+            and _is_fp8_cutlass_moe_supported()
+        ):
+            override_server_args(
+                server_args,
+                "sglang-omni-qwen3-backend-policy",
+                moe_runner_backend="cutlass",
+            )
+            moe_runner_backend = server_args.moe_runner_backend
+
+        if (
+            is_qwen3_omni_arch
+            and effective_quantization == "fp8"
+            and has_moe
+            and moe_runner_backend == "cutlass"
+        ):
+            if not has_native_fp8_block_quant:
+                raise ValueError(
+                    "Qwen3-Omni FP8 CUTLASS MoE requires a native serialized "
+                    "block-FP8 checkpoint with weight_block_size."
+                )
+
+        if (
+            is_qwen3_omni_arch
+            and effective_quantization == "fp8"
+            and moe_runner_backend == "flashinfer_cutlass"
+        ):
+            raise ValueError(
+                "Qwen3-Omni native FP8 checkpoints cannot use "
+                "moe_runner_backend='flashinfer_cutlass'. Leave the backend as "
+                "'auto' so Omni selects a native-FP8-compatible MoE runner."
+            )
+
+        fp8_gemm_backend = normalize_quantization(server_args.fp8_gemm_runner_backend)
+        if (
+            model_arch_override == "Qwen3OmniTalker"
+            and effective_quantization == "fp8"
+            and has_native_fp8_block_quant
+            and fp8_gemm_backend in (None, "auto")
+        ):
+            # Projected talker prefill has request-dependent FP8 dense GEMM shapes
+            # outside decode CUDA graph replay; DeepGEMM can otherwise JIT there.
+            override_server_args(
+                server_args,
+                "sglang-omni-qwen3-backend-policy",
+                fp8_gemm_runner_backend="triton",
+            )
+            fp8_gemm_backend = server_args.fp8_gemm_runner_backend
+
+        server_quantization = server_args.quantization
+        logger.info(
+            f"Configured SGLang backend policy: arch={model_arch_override} "
+            f"effective_quantization={effective_quantization} "
+            f"server_quantization={server_quantization} "
+            f"moe_runner_backend={moe_runner_backend} "
+            f"fp8_gemm_backend={fp8_gemm_backend}"
+        )
+        return effective_quantization
 
 
 class ROCMOmniPlatform(RocmDeviceMixin, CUDAOmniPlatform):
