+++ sglang_omni/platforms/interface.py
@@ -7,7 +7,10 @@ from typing import TYPE_CHECKING
 
 from sglang.srt.platforms.device_mixin import DeviceMixin
 
+from sglang_omni.utils.misc import normalize_quantization
+
 if TYPE_CHECKING:
+    from sglang_omni.comm.data_ref import TransportKind
     from sglang_omni.pipeline.stage_workers import StageLaunchConfig
 
 
@@ -21,3 +24,26 @@ class OmniPlatform(DeviceMixin):
     ) -> dict[str, str]:
         """Return per-process environment overrides needed before child startup."""
         return {}
+
+    def get_intra_node_transport(self) -> TransportKind:
+        from sglang_omni.comm.data_ref import TransportKind
+
+        return TransportKind.SHM
+
+    def get_fused_qk_norm_rope(self):
+        """Get the fused QK norm RoPE kernel if available, else return None."""
+        return None
+
+    def apply_model_worker_backend_policy(
+        self,
+        server_args: ServerArgs,
+        model_config: ModelConfig,
+        model_arch_override: str | None,
+    ) -> str | None:
+        """Apply Omni backend policy after checkpoint quantization is known."""
+
+        effective_quantization = normalize_quantization(model_config.quantization)
+        server_quantization = normalize_quantization(server_args.quantization)
+        if server_quantization is not None:
+            effective_quantization = server_quantization
+        return effective_quantization
