+++ sglang_omni/utils/misc.py
@@ -95,3 +95,9 @@ def broadcast_pyobj(
         serialized_data = bytes(tensor_data.cpu().numpy())
         data = pickle.loads(serialized_data)
         return data
+
+
+def normalize_quantization(value: object) -> str | None:
+    if value is None:
+        return None
+    return str(value).lower()
