+++ tests/unit_test/qwen3_omni/test_fp8_backend_config.py
@@ -9,6 +9,7 @@ from types import ModuleType, SimpleNamespace
 import pytest
 
 from sglang_omni.model_runner import model_worker
+from sglang_omni.platforms import cuda
 from tests.unit_test.fakes import FakeServerArgs
 
 
@@ -357,7 +358,7 @@ def test_model_worker_backend_policy_precedence(
         "_is_fp8_cutlass_moe_supported",
         lambda: case.cutlass_supported,
     )
-    monkeypatch.setattr(model_worker, "_is_h20_device", lambda: False)
+    monkeypatch.setattr(cuda, "_is_h20_device", lambda: False)
     server_args = _server_args(
         quantization=case.server_quantization,
         moe_runner_backend=case.initial_moe_backend,
@@ -372,14 +373,14 @@ def test_model_worker_backend_policy_precedence(
 
     if case.error_match:
         with pytest.raises(ValueError, match=case.error_match):
-            model_worker._apply_model_worker_backend_policy(
+            current_platform.apply_model_worker_backend_policy(
                 server_args,
                 model_config,
                 case.model_arch_override,
             )
         return
 
-    effective_quantization = model_worker._apply_model_worker_backend_policy(
+    effective_quantization = current_platform.apply_model_worker_backend_policy(
         server_args,
         model_config,
         case.model_arch_override,
@@ -412,7 +413,7 @@ def test_model_worker_backend_policy_uses_strict_server_args_override(
         has_moe=True,
     )
 
-    effective_quantization = model_worker._apply_model_worker_backend_policy(
+    effective_quantization = current_platform.apply_model_worker_backend_policy(
         server_args,
         model_config,
         "Qwen3OmniTalker",
@@ -525,11 +526,11 @@ class FullConfigureBackendPolicyCase:
 
 
 # Test cases covering the ordering issue: the Omni quantization adapters
-# run BEFORE _apply_model_worker_backend_policy(), so only Talker FP8
+# run BEFORE apply_model_worker_backend_policy(), so only Talker FP8
 # with native block quant should get triton GEMM; Thinker and non-Qwen
 # should preserve auto. The adapters are a no-op for FP8 (they only
 # normalize stage-local names for methods like AutoRound), so all FP8
-# backend policy stays owned by _apply_model_worker_backend_policy().
+# backend policy stays owned by apply_model_worker_backend_policy().
 CONFIGURE_BACKEND_POLICY_CASES = [
     FullConfigureBackendPolicyCase(
         name="talker_fp8_auto_gemm_becomes_triton",
@@ -689,7 +690,7 @@ def test_configure_backend_policy_fp8_gemm_ordering(
 
     The ordering in _configure_backend_policy() is:
         1. _apply_omni_quantization_adapters()  (no-op for FP8)
-        2. _apply_model_worker_backend_policy()
+        2. apply_model_worker_backend_policy()
 
     Only step 2 (arch-aware) should set fp8_gemm_runner_backend="triton"
     for Talker FP8. Step 1 must NOT touch FP8 backend selection.
@@ -712,7 +713,7 @@ def test_configure_backend_policy_fp8_gemm_ordering(
     )
 
     # Patch _is_h20_device so we get deterministic BF16 policy.
-    monkeypatch.setattr(model_worker, "_is_h20_device", lambda: False)
+    monkeypatch.setattr(cuda, "_is_h20_device", lambda: False)
 
     # Build mock model config matching the shape ModelConfig expects.
     quant_config_in = (
@@ -754,9 +755,9 @@ def test_configure_backend_policy_fp8_gemm_ordering(
     # ordering in _configure_backend_policy().
     model_worker._apply_omni_quantization_adapters(model_config)
 
-    # Step 2: run the REAL _apply_model_worker_backend_policy().
+    # Step 2: run the REAL apply_model_worker_backend_policy().
     # This is the arch-aware step that sets Talker FP8 Triton.
-    _ = model_worker._apply_model_worker_backend_policy(
+    _ = current_platform.apply_model_worker_backend_policy(
         server_args,
         model_config,
         case.model_arch_override,
